Move ignoreCheckTo init part
